pyproject.toml
[project]
name = "paperhunt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.1.3",
    "google-generativeai>=0.8.4",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "flake8>=7.1.1",
    "isort>=5.13.2",
]
